fix(sonar): fix some sonar issues
