+++ ide_extension/intellij/src/main/java/co/com/bancolombia/devsecopsenginetools/actions/ScanIacAction.java
@@ -7,15 +7,25 @@ import com.intellij.openapi.actionSystem.AnAction;
 import com.intellij.openapi.actionSystem.AnActionEvent;
 import com.intellij.openapi.progress.ProgressManager;
 import com.intellij.openapi.project.Project;
+import com.intellij.openapi.util.NlsActions;
 import lombok.SneakyThrows;
+import org.jetbrains.annotations.Nullable;
+
+import javax.swing.*;
 
 public class ScanIacAction extends AnAction {
 
     public ScanIacAction() {
+        super();
     }
 
-    public ScanIacAction(Boolean isRerun) {
-        super("Scan IaC", "Run iac scan", AllIcons.Actions.Execute);
+    public ScanIacAction(@Nullable @NlsActions.ActionText String text,
+                         @Nullable @NlsActions.ActionDescription String description, @Nullable Icon icon) {
+        super(text, description, icon);
+    }
+
+    public static ScanIacAction forUI() {
+        return new ScanIacAction("Scan IaC", "Run iac scan", AllIcons.Actions.Execute);
     }
 
     @SneakyThrows
