+++ ide_extension/intellij/src/main/java/co/com/bancolombia/devsecopsenginetools/utils/Constants.java
@@ -1,0 +1,10 @@
+package co.com.bancolombia.devsecopsenginetools.utils;
+
+import lombok.experimental.UtilityClass;
+
+@UtilityClass
+public class Constants {
+    public static final String AZURE_CREDENTIALS = "azure";
+    public static final String DEFAULT_URL_PATTERN = "{...}";
+    public static final String DEFAULT_PATTERN = "#{...}#";
+}
