+++ ide_extension/intellij/src/main/java/co/com/bancolombia/devsecopsenginetools/utils/variables/azure/exceptions/VariableGroupNotFoundException.java
@@ -1,0 +1,7 @@
+package co.com.bancolombia.devsecopsenginetools.utils.variables.azure.exceptions;
+
+public class VariableGroupNotFoundException extends RuntimeException {
+    public VariableGroupNotFoundException(String message) {
+        super(message);
+    }
+}
